Extract and update DaMaRiS to run with Geant4 11.2
Remove use of G4PhysicsModelCalaogue to generate unique ID for BreakMolecule, instead an integer ID is now generated
in DrDefinitions. This ID is used to associate an AuxilliaryTrackInformation to the Track of each molecule.

Should explore if we can/should still use G4PhysicsModelCatalogue in future to ensure that DaMaRiS doesn't clash with
other uses of AuxTrackInfo.

Additionally have added output of RawData.out which is the preferred method to generate data with DaMaRiS (over
PerRunResultsX.out). This just outputs the integer values of molecules present in the simulation at each time point
so the user can do theiur own post processing/calculations on the raw data.